+++ Datawarehouse/Database Staging/Map/Tables/Dim_Produkt.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [Map].[Dim_Produkt]
 (   
-    [Ekey_Produkt]          BIGINT         NOT NULL IDENTITY (1,1),
-	[Product_Id]            BIGINT         NOT NULL,
+    [Ekey_Produkt]          INT           NOT NULL IDENTITY (1,1),
+	[Product_Id]            INT           NOT NULL,
 
     /* Metadata */
 	[Meta_IsDeleted]		BIT			   NOT NULL,
